refactor: change paths to animations and services
